Parsing, linting and refactoring

+++ apps/experiment/management/commands/delete_fake_data.py
@@ -1,7 +1,8 @@
 import os
 
+from django.contrib.auth import get_user_model
 from django.core.management.base import BaseCommand
-from django.contrib.auth.models import User
+
 from experiment.models import Experiment
 
 
@@ -11,10 +12,10 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         self.stdout.write("Deleting data...")
         Experiment.objects.all().delete()
-        User.objects.filter(is_superuser=False).delete()
+        get_user_model().objects.filter(is_superuser=False).delete()
         self.stdout.write("...finished")
 
         message_folder = os.path.join(os.getcwd(), 'messages')
-        for f in os.listdir(message_folder):
-            os.remove(os.path.join(message_folder, f))
-
+        if os.path.isdir(message_folder):
+            for f in os.listdir(message_folder):
+                os.remove(os.path.join(message_folder, f))
